requirements.txt
langchain
langchain_community
langchain-google-genai
langchain-huggingface
langchain_experimental
langchain_chroma
langchainhub
pypdf
ipykernel
rapidocr-onnxruntime
torch
transformers
sentence-transformers
python-dotenv
streamlit
unstructured
google-search-results
openai
arxiv